+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## What's Changed
 * `amount` and `maxAmount` fields from `StackSlot<T>` and `LazyStackSlot<T>` are now private
+* `StackSlot<T>.Content` is now an `IEnumerable<T>`
 
 ## What's Next
 * Improves in Container classes to implement from interfaces from `System.Collections.Generic`