+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# v0.20.0
+
+## What's Added
+* 
+
+## What's Changed
+* 
+
+## What's Fixed
+* 
+
+## What's Next
+* Improves in Container classes to implement from interfaces from `System.Collections.Generic`
+* The protected field `slots` in Container classes can be changed to an protected property.
+
+**Full Changelog**: https://github.com/The-Chest/TheChest.Core/compare/v0.19.0...v0.20.0
+
 # v0.19.0
 
 ## What's Changed